Fix S3 access error bug on landing page demo
Previously when a user reloaded the page, create_new_temp_file was set to '',
which meant that the editor would try to load the temp file from S3. However,
if the user hadn't saved the editor at that point, the temp file wouldn't exist
in S3. Now create_new_temp_file is only set to '' when a temp file has been
saved to S3.